+++ frontend/src/components/NavBar.tsx
@@ -1,13 +1,13 @@
-import { path } from "@/lib/constants";
+import { PATH } from "@/lib/constants";
 import { NavBarLink } from "./NavBarLink";
 
 function NavBar() {
   return (
     <nav className="flex gap-2 p-2">
-      <NavBarLink href={path.ROOT}>Главная</NavBarLink>
-      <NavBarLink href={path.EXPRESSIONS}>Примеры</NavBarLink>
-      <NavBarLink href={path.WORD_GAME}>Слово</NavBarLink>
-      <NavBarLink href={path.MATCHES}>Соответствия</NavBarLink>
+      <NavBarLink href={PATH.ROOT}>Главная</NavBarLink>
+      <NavBarLink href={PATH.EXPRESSIONS}>Примеры</NavBarLink>
+      <NavBarLink href={PATH.WORD_GAME}>Слово</NavBarLink>
+      <NavBarLink href={PATH.MATCHES}>Соответствия</NavBarLink>
     </nav>
   );
 }