+++ frontend/src/lib/api.ts
@@ -1,12 +1,13 @@
-import { HTTP_TOO_MANY_REQUESTS } from "./constants";
+import { BACKEND_URL, HTTP_TOO_MANY_REQUESTS } from "./constants";
+import { Path } from "./types";
 import { isBaseAPIErrorResponse } from "./utils";
 
 async function apiFetch(
-  path: string,
+  path: Path,
   init?: RequestInit,
   errorMessage?: string,
 ): Promise<any> {
-  const response = await fetch(path, { ...init });
+  const response = await fetch(`${BACKEND_URL}${path}`, { ...init });
   const data = await response.json();
 
   if (response.ok) return data;