+++ CHANGELOG.md
@@ -4,6 +4,22 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+Version numbers are stored in: `frontend/package.json`, `backend/src/config/constants.py` and `backend/pyproject.toml`.
+
+## [3.1.0] - December 5, 2025
+
+### Added
+
+- Add 2-4 grades images to matches page
+
+### Changed
+
+- Transfer matches images to CDN
+
+### Fixed
+
+- Fix responsiveness issues in matches countdown
+
 ## [3.0.3] - December 5, 2025
 
 ### Changed